refactor: derive section hrefs from COLLECTION_SEGMENTS
The rebase left two collection→URL sources: COLLECTION_SEGMENTS in
utils/collections.ts (from the drafts-permalink work) and the SECTIONS
hrefs in content.config.ts. Move the segment map into content.config
next to the collections it describes and build the hrefs from it, so
renaming a segment cannot desync the sitemap, nav, or section pages.
getEntryPath stays in utils/collections.
